Changes
Improved image quality for instruction images
Udpdated the commenting and logic of OSARI_functions.py's 'calculateStopTime' function. Correct and incorrect are now referred to as 2s and 0s (consistent with OSARI.py). "Signal: (thisTrial['Signal'] in OSARI.py) was added as an argument to the function in order to make the changes only occur following stop trials.
practiceMixedTrials.xlsx has stop and go's interspersed (0,1,0,1,0,1 and so on). This was done because those trials are read in sequentially.
testBlocks.xlsx formatted so that fixedStopTime column are all "-".