OSARI powerpoint stimuli
Powerpoint slides containing the stimuli are now available to users if they wish to change the content of the instructions (e.g., users may wish to present instructions in another language). Given that the instructions are presented as .jpeg images, they could not be easily edited by the user. The provision of the powerpoint slides allows users to simply edit the text in the text boxes. Users will still have to save the stimuli as .jpeg files, and instructions to do this are provided as the first slide of the powerpoint.